test: split coordinator coverage tests by behavior
